chore: #94 move block_id to separate file
